// types/constants.ts
export enum Month {
  JANUARY = "jan",
  FEBRUARY = "feb",
  MARCH = "mar",
  APRIL = "apr",
  MAY = "may",
  JUNE = "jun",
  JULY = "jul",
  AUGUST = "aug",
  SEPTEMBER = "sep",
  OCTOBER = "oct",
  NOVEMBER = "nov",
  DECEMBER = "dec",
}

export enum MonthName {
  JANUARY = "Январь",
  FEBRUARY = "Февраль",
  MARCH = "Март",
  APRIL = "Апрель",
  MAY = "Май",
  JUNE = "Июнь",
  JULY = "Июль",
  AUGUST = "Август",
  SEPTEMBER = "Сентябрь",
  OCTOBER = "Октябрь",
  NOVEMBER = "Ноябрь",
  DECEMBER = "Декабрь",
}

export enum Kind {
  COSTS = "costs",
  INCOME = "income",
}

export enum RuKind {
  COSTS = "Расходы",
  INCOME = "Доходы",
}

export enum Category {
  PRODUCTS = "products",
  ENTERTAINMENT = "entertainment",
  TRANSPORT = "transport",
  HEALTH = "health",
  HOME = "home",
  EDUCATION = "education",
  FITNESS = "fitness",
  TAXES = "taxes",
  SALARY = "salary",
  REWARD = "reward",
  PRESENT = "present",
  SALES = "sales",
  OTHER = "other",
}

export enum RuCategory {
  PRODUCTS = "Продукты",
  ENTERTAINMENT = "Развлечения",
  TRANSPORT = "Транспорт",
  HEALTH = "Здоровье",
  HOME = "ЖКХ",
  EDUCATION = "Обучение",
  FITNESS = "Фитнес",
  TAXES = "Налоги",
  SALARY = "Зарплата",
  REWARD = "Награда",
  PRESENT = "Подарок",
  SALES = "Продажа",
  OTHER = "Другое",
}

export enum CostsCategories {
  PRODUCTS = "products",
  ENTERTAINMENT = "entertainment",
  TRANSPORT = "transport",
  HEALTH = "health",
  HOME = "home",
  EDUCATION = "education",
  FITNESS = "fitness",
  TAXES = "taxes",
}

export enum IncomeCategories {
  SALARY = "Зарплата",
  REWARD = "Награда",
  PRESENT = "Подарок",
  SALES = "Продажа",
  OTHER = "Другое",
}